Add files via upload
Initial commit. The HDL codes attached to this commit have been tested. 'tb_processor' is the testbench module that must be used for simulation. 'simpleRISC.v' the top level file contains dummy data and will be used for connecting the FPGA ports/pins with the 'processor' module counterparts in the following commits. To setup the Quartus project use the simpleRISC.tcl script.